+++ apps/web/package.json
@@ -56,6 +56,7 @@
     "require-in-the-middle": "^7.5.2",
     "superjson": "^2.0.0",
     "unified": "^11.0.5",
+    "use-debounce": "^10.0.5",
     "zod": "^3.25.47"
   },
   "devDependencies": {
