+++ apps/web/src/app/artikler/queries.ts
@@ -1,0 +1,22 @@
+import { useTRPC } from "@/utils/trpc/client"
+import type { ArticleFilterQuery } from "@dotkomonline/types"
+
+import { useInfiniteQuery, useQuery } from "@tanstack/react-query"
+
+export const useArticleAllQuery = () => {
+  const trpc = useTRPC()
+  return useQuery(trpc.article.all.queryOptions())
+}
+
+export const useArticleFilterQuery = (filters: ArticleFilterQuery) => {
+  const trpc = useTRPC()
+
+  const { data, ...query } = useInfiniteQuery({
+    ...trpc.article.findArticles.infiniteQueryOptions({
+      filters
+    }),
+    getNextPageParam: (lastPage) => lastPage.nextCursor,
+  })
+
+  return { data, ...query }
+}
